Changed name of cgpa and sgpa, alson added a gpa page which conditional renders sgpa and cgpa pages. also changed cgpa calc into gpa calc
